+++ src/Shape/triangle.hpp
@@ -10,7 +10,7 @@ public:
 
 	Triangle(const Vec& p1,const Vec& p2,const Vec& p3, const Vec& e, const Vec& c, Refl_t refl) ;
 	virtual ~Triangle();
-	virtual bool intersect(const Ray &pixelCol, double &eps);
+	virtual bool intersect(const Ray &pixelCol, float &eps);
 	virtual Vec getNorm(Vec x);
 };
 #endif
